feat: offer second processor speeds that are not multiples of 3MHz
The slider now picks from the speeds that existed or are worth having: 3MHz
for the external 6502, 4MHz for the Master Turbo's 65C102, then 6, 8, 12 and
24MHz. tubeCpuMultiplier is parsed as a float so integer links keep their
meaning, and the label leads with the speed.

Fixes #733

+++ src/main.js
@@ -138,7 +138,7 @@ const paramTypes = {
     audiofilterfreq: ParamTypes.FLOAT,
     audiofilterq: ParamTypes.FLOAT,
     cpuMultiplier: ParamTypes.FLOAT,
-    tubeCpuMultiplier: ParamTypes.INT,
+    tubeCpuMultiplier: ParamTypes.FLOAT,
     microphoneChannel: ParamTypes.INT,
 
     // String parameters (these are the default but listed for clarity)
